REL-325028 Fix configuration issues

+++ Source/Relativity.Sync/Configuration/IDataSourceSnapshotConfiguration.cs
@@ -19,6 +19,6 @@ namespace Relativity.Sync.Configuration
 
 		bool IsSnapshotCreated { get; }
 
-		Task SetSnapshotDataAsync(Guid runId, long totalRecordsCount);
+		Task SetSnapshotDataAsync(Guid runId, int totalRecordsCount);
 	}
 }
