+++ Source/Relativity.Sync/Executors/DataSourceSnapshotExecutor.cs
@@ -66,7 +66,7 @@ namespace Relativity.Sync.Executors
 
 			//ExportInitializationResult provide list of fields with order they will be returned when retrieving metadata
 			//however, order is the same as order of fields in QueryRequest when they are provided explicitly
-			await configuration.SetSnapshotDataAsync(results.RunID, results.RecordCount).ConfigureAwait(false);
+			await configuration.SetSnapshotDataAsync(results.RunID, (int)results.RecordCount).ConfigureAwait(false);
 			return ExecutionResult.Success();
 		}
 	}
